+++ examples/provider/provider.tf
@@ -1,5 +1,7 @@
 provider "appwrite" {
-  endpoint   = "https://cloud.appwrite.io/v1"
-  project_id = "project-id"
-  api_key    = "api-key"
+  endpoint             = "https://cloud.appwrite.io/v1"
+  project_id           = "project-id"
+  organization_id      = "organization-id"
+  api_key              = "project-api-key"
+  organization_api_key = "organization-api-key"
 }