+++ examples/resources/appwrite_project/resource.tf
@@ -1,0 +1,5 @@
+resource "appwrite_project" "example" {
+  id     = "example-project"
+  name   = "Example Project"
+  region = "fra"
+}