+++ examples/resources/appwrite_project_key/import.sh
@@ -1,0 +1,1 @@
+terraform import appwrite_project_key.example organization-id/project-id/deployments